# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Project Talker Manager",
    "version": "1.0.0"
  },
...
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "/login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "email": "matheus@gmail.com",
                  "password": "helloworld"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/talker": {
      "get": {
        "tags": [
          "Talker"
        ],
        "summary": "/talker",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "Talker"
        ],
        "summary": "/talker",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Danielle Santos",
                  "age": 56,
                  "talk": {
                    "watchedAt": "22/10/2019",
                    "rate": 5
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "example": "5c00549a8c52aa65"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}